requirements.txt
flask
waitress
pillow
pytorch
transformers